requirements.txt

pyyaml
geopandas
osmnx
pytest
pandas
matplotlib
contextily
numpy
imageio
shapely